move TagsQuery stuff into its own folder
