+++ src/exampleTypes/TagsQuery/TagPopover.js
@@ -2,9 +2,9 @@ import React from 'react'
 import _ from 'lodash/fp'
 import F from 'futil-js'
 import { observer } from 'mobx-react'
-import TagsJoinPicker from './TagsJoinPicker'
-import { withTheme } from '../utils/theme'
-import { getTag } from '../utils/tagsQuery'
+import TagsJoinPicker from '../TagsJoinPicker'
+import { withTheme } from '../../utils/theme'
+import { getTag } from './utils'
 
 let TagQueryPopover = ({
   tag,
