+++ src/exampleTypes/index.js
@@ -20,4 +20,3 @@ export TermsStats from './TermsStats'
 export TermsStatsTable from './TermsStatsTable'
 export Text from './Text'
 export TypeMap from './TypeMap'
-export TagQueryPopover from './TagQueryPopover'
